+++ src/runmd/cli.py
@@ -25,6 +25,7 @@ from pathlib import Path
 from typing import Optional
 
 from . import __version__
+from .commands import create_parser
 from .config import get_configuration
 from .history import load_history, print_history, update_history, write_history
 from .process import list_command, process_markdown_files, run_command, show_command
@@ -33,113 +34,6 @@ RUNCMD = "run"
 SHOWCMD = "show"
 LISTCMD = "list"
 HISTCMD = "hist"
-
-
-def cliargs() -> argparse.ArgumentParser:
-    """
-    Create and return the argument parser for the CLI tool.
-
-    This function defines the structure of the available commands, including their arguments,
-    and returns the configured argument parser.
-
-    Returns:
-        argparse.ArgumentParser: The argument parser configured with all subcommands.
-    """
-    # Create the top-level parser
-    parser = argparse.ArgumentParser(
-        prog="runmd",
-        description="A tool to manage and process code blocks in Markdown files.",
-    )
-
-    # Argument to show current version
-    parser.add_argument(
-        "--version", action="version", version=f"%(prog)s {__version__}"
-    )
-
-    # Create subparsers for different commands
-    subparsers = parser.add_subparsers(dest="command", help="Available commands")
-
-    # Common arguments for all subparsers
-    common_parser = argparse.ArgumentParser(add_help=False)
-    common_parser.add_argument(
-        "-f",
-        "--file",
-        nargs="?",
-        default=None,
-        help="Path to the markdown file to process. If not provided, uses the default file from config.",
-    )
-
-    # Subparser for the 'run' command
-    run_parser = subparsers.add_parser(
-        RUNCMD,
-        help="Run code blocks in the source file",
-        parents=[common_parser],
-    )
-    run_parser.add_argument(
-        "blockname",
-        nargs="?",
-        default=None,
-        help='Name of the code block to run or "all" to run all blocks',
-    )
-    run_parser.add_argument(
-        "-t",
-        "--tag",
-        nargs="?",
-        default=None,
-        help="Execute all code blocks with this tag",
-    )
-    run_parser.add_argument(
-        "--env",
-        nargs="*",
-        default=[],
-        help="Environment variables to set during execution (e.g., VAR=value)",
-    )
-
-    # Subparser for the 'show' command
-    show_parser = subparsers.add_parser(
-        SHOWCMD,
-        help="Show code blocks from the source file",
-        parents=[common_parser],
-    )
-    show_parser.add_argument(
-        "blockname",
-        nargs="?",
-        help="Name of the code block to show",
-    )
-
-    # Subparser for the 'list' command
-    list_parser = subparsers.add_parser(
-        LISTCMD,
-        help="List code blocks in the source file",
-        parents=[common_parser],
-    )
-    list_parser.add_argument(
-        "-t",
-        "--tag",
-        nargs="?",
-        default=None,
-        help="Optional tag to filter the list of code blocks",
-    )
-
-    # Subparser for the 'hist' command
-    hist_parser = subparsers.add_parser(
-        HISTCMD,
-        help="Display or manage the runmd command history",
-    )
-    hist_group = hist_parser.add_mutually_exclusive_group()
-    hist_group.add_argument(
-        "id",
-        nargs="?",
-        default=None,
-        help="ID of the command to run from history",
-    )
-    hist_group.add_argument(
-        "--clear",
-        action="store_true",
-        help="Clear the history list",
-    )
-
-    return parser
 
 
 def execute_command(
@@ -164,7 +58,7 @@ def execute_command(
         if args.id:
             oldcmd = history[int(args.id)]["command"].split(" ")
             # Reparse
-            parser = cliargs()
+            parser = create_parser()
             args = parser.parse_args(oldcmd[1:])
         elif args.clear:
             history = []
@@ -210,7 +104,7 @@ def main(command_line: Optional[list[str]] = None) -> None:
     config = get_configuration()
 
     # Parse the command-line arguments
-    parser = cliargs()
+    parser = create_parser()  # cliargs()
     args = parser.parse_args(command_line)
 
     # Handle case where no command is provided